Started the in-game repositry
Added a skeleton code block for the in-game calander system which will
allow the game torack what day and week it is. this is being used to as
the backbone of thew game to ensure the plot playes out as intended.

It is built using a switch statment qhich quires the status of an enum
property for week day.